requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
httpx>=0.24,<0.26
aiohttp==3.9.1
redis==5.0.1
supabase==2.0.3
apscheduler==3.10.4
pydantic==2.6.1
python-jose==3.3.0
python-multipart==0.0.9
rich==13.7.0
psycopg2-binary==2.9.9
celery==5.3.6
flower==2.0.1
sendgrid==6.10.0
gunicorn==20.1.0 